made admin access a little better

+++ client/src/AdminPage/AdminPage.jsx
@@ -15,32 +15,45 @@ class AdminPage extends React.Component {
 
     render() {
         const { user, users } = this.props;
-        return (
-            <div className="col-md-6 col-md-offset-3">
-                <h1>Hi {user.firstName}!</h1>
-                <p>You're logged in with React!!</p>
-                <h3>All registered users:</h3>
-                {users.loading && <em>Loading users...</em>}
-                {users.error && <span className="text-danger">ERROR: {users.error}</span>}
-                {users.items &&
-                    <ul>
-                        {users.items.map((user, index) =>
-                            <li key={user.id}>
-                                {user.firstName + ' ' + user.lastName}
-                                {
-                                    user.deleting ? <em> - Deleting...</em>
-                                    : user.deleteError ? <span className="text-danger"> - ERROR: {user.deleteError}</span>
-                                    : <span> - <a onClick={this.handleDeleteUser(user.id)}>Delete</a></span>
-                                }
-                            </li>
-                        )}
-                    </ul>
-                }
-                <p>
+        if (user.role === 'Admin')
+        {
+            return (
+                <div className="col-md-6 col-md-offset-3">
+                    <h1>Hi {user.firstName}!</h1>
+                    <p>You're logged in with React!!</p>
+                    <h3>All registered users:</h3>
+                    {users.loading && <em>Loading users...</em>}
+                    {users.error && <span className="text-danger">ERROR: {users.error}</span>}
+                    {users.items &&
+                        <ul>
+                            {users.items.map((user, index) =>
+                                <li key={user.id}>
+                                    {user.firstName + ' ' + user.lastName}
+                                    {
+                                        user.deleting ? <em> - Deleting...</em>
+                                        : user.deleteError ? <span className="text-danger"> - ERROR: {user.deleteError}</span>
+                                        : <span> - <a onClick={this.handleDeleteUser(user.id)}>Delete</a></span>
+                                    }
+                                </li>
+                            )}
+                        </ul>
+                    }
+                    <p>
+                        <Link to="/login">Logout</Link>
+                    </p>
+                </div>
+            );
+        } else if (user.role === 'User') {
+            return (
+                <div className="col-md-6 col-md-offset-3">
+                    <h1>Hi {user.firstName}!</h1>
+                    <p>This is the admin page, and only accessible to admins</p>
+                    <p>
                     <Link to="/login">Logout</Link>
-                </p>
-            </div>
-        );
+                    </p>
+                </div>
+            )
+        }
     }
 }
